+++ src/pytest_workflow/plugin.py
@@ -132,8 +132,7 @@ def pytest_collection(session):
 def pytest_runtestloop(session):
     """This runs after collection, but before the tests."""
     session.config.workflow_queue.process(
-        number_of_threads=session.config.getoption("workflow_threads"),
-        save_logs=session.config.getoption("keep_workflow_wd")
+        session.config.getoption("workflow_threads")
     )
 
 
@@ -240,14 +239,14 @@ class WorkflowTestsCollector(pytest.Collector):
 
         tests += [ContentTestCollector(
             name="stdout", parent=self,
-            content_generator=workflow.stdout_lines,
+            filepath=workflow.stdout_file,
             content_test=self.workflow_test.stdout,
             workflow=workflow,
             content_name="'{0}': stdout".format(self.workflow_test.name))]
 
         tests += [ContentTestCollector(
             name="stderr", parent=self,
-            content_generator=workflow.stderr_lines,
+            filepath=workflow.stderr_file,
             content_test=self.workflow_test.stderr,
             workflow=workflow,
             content_name="'{0}': stderr".format(self.workflow_test.name))]